# perf: optimize performance of waybar cpuinfo and gpuinfo modules