--- a/Configs/.config/hypr/scripts/gpuinfo.sh
+++ b/Configs/.config/hypr/scripts/gpuinfo.sh
@@ -102,57 +102,53 @@ sed -i "s/^#${next_prioGPU}/${next_prioGPU}/" "${gpuQ}" # Uncomment the next pri
 sed -i "s/prioGPU=${prioGPU}/prioGPU=${next_prioGPU}/" "${gpuQ}" # Update the prioGPU in the file
 }
 
-# Thee shalt find the greatest one,
-# He who not more than the chosen one
-map_floor() {
-
-    # From the depths of the string, words arise,
-    # Keys in pairs, a treasure in disguise.
-    IFS=', ' read -r -a pairs <<< "$1"
-
-    # If the final token stands alone and bold,
-    # Declare it the default, its worth untold.
-    if [[ ${pairs[-1]} != *":"* ]]; then
-        def_val="${pairs[-1]}"
-        unset 'pairs[${#pairs[@]}-1]'
-    fi
-
-    # Scans the map, a peak it seeks,
-    # The highest passed, the value speaks.
-    for pair in "${pairs[@]}"; do
-        IFS=':' read -r key value <<< "$pair"
-        
-        # Behold! Thou holds the secrets they seek,
-        # Declare it and silence the whispers unique.
-        if awk -v num="$2" -v k="$key" 'BEGIN { exit !(num > k) }'; then
-            echo "$value"
-            return
-        fi
-    done
-
-    # On this lonely shore, where silence dwells
-    # Even the waves, echoes words unheard
-    [ -n "$def_val" ] && echo $def_val || echo " "
-}
-
-# generate emoji and icon based on temperature and utilization
 get_icons() {
-    # key-value pairs of temperature and utilization levels
-    temp_lv="85:&🌋, 65:&🔥, 45:&☁️, &❄️"
+
+    # Thee shalt find the greatest one,
+    # He who not more than the chosen one
+    map_floor() {
+
+        # From the depths of the string, words arise,
+        # Keys in pairs, a treasure in disguise.
+        IFS=', ' read -r -a pairs <<< "$1"
+
+        # If the final token stands alone and bold,
+        # Declare it the default, its worth untold.
+        if [[ ${pairs[-1]} != *":"* ]]; then
+            def_val="${pairs[-1]}"
+            unset 'pairs[${#pairs[@]}-1]'
+        fi
+
+        # Scans the map, a peak it seeks,
+        # The highest passed, the value speaks.
+        for pair in "${pairs[@]}"; do
+            IFS=':' read -r key value <<< "$pair"
+            
+            # Behold! Thou holds the secrets they seek,
+            # Declare it and silence the whispers unique.
+            if [ ${2%%.*} -gt $key ]; then
+                echo "$value"
+                return
+            fi
+        done
+
+        # On this lonely shore, where silence dwells
+        # Even the waves, echoes words unheard
+        [ -n "$def_val" ] && echo $def_val || echo " "
+    }
+
+    temp_lv="85:🌋, 65:🔥, 45:☁️, ❄️"
     util_lv="90:, 60:󰓅, 30:󰾅, 󰾆" 
 
-    # return comma seperated emojis/icons 
-    icons=$(map_floor "$temp_lv" $1 | sed "s/&/,/")
-    icons="$icons,$(map_floor "$util_lv" $2)"
-    echo $icons
+    echo "$(map_floor "$util_lv" $2)$(map_floor "$temp_lv" $1)"
 }
 
 generate_json() {
   # get emoji and icon based on temperature and utilization
   icons=$(get_icons "$temperature" "$utilization")
-  thermo=$(echo $icons | awk -F, '{print $1}')
-  emoji=$(echo $icons | awk -F, '{print $2}')
-  speedo=$(echo $icons | awk -F, '{print $3}')
+  speedo=$(echo ${icons:0:1})
+  thermo=$(echo ${icons:1:1})
+  emoji=$(echo ${icons:2})
 
   # emoji=$(get_temperature_emoji "${temperature}")
   local json="{\"text\":\"${thermo} ${temperature}°C\", \"tooltip\":\"${primary_gpu}\n${thermo} Temperature: ${temperature}°C ${emoji}"