--- a/Configs/.config/hypr/scripts/gpuinfo.sh
+++ b/Configs/.config/hypr/scripts/gpuinfo.sh
@@ -4,6 +4,8 @@
 gpuQ="/tmp/hyprdots-${UID}-gpuinfo-query"
 tired=false
 [[ " $* " =~ " tired " ]] && ! grep -q "tired" "${gpuQ}" && echo "tired=true" >>"${gpuQ}"
+[[ " $* " =~ " no_emoji " ]] && ! grep -q "NO_EMOJI" "${gpuQ}" && echo "NO_EMOJI=1" >>"${gpuQ}"
+
 if [[ ! " $* " =~ " startup " ]]; then
    gpuQ="${gpuQ}$2"
 fi
@@ -135,14 +137,6 @@ map_floor() {
     [ -n "$def_val" ] && echo $def_val || echo " "
 }
 
-# Define glyphs
-if [[ $NO_EMOJI -eq 1 ]]; then
-    temp_lv="85:, 65:, 45:☁, ❄"
-else
-    temp_lv="85:🌋, 65:🔥, 45:☁️, ❄️"
-fi
-util_lv="90:, 60:󰓅, 30:󰾅, 󰾆" 
-
 generate_json() {
   # Generate glyphs
   icons=$(echo "$(map_floor "$util_lv" $utilization)$(map_floor "$temp_lv" $temperature)")
@@ -267,12 +261,21 @@ Avalable GPU: ${gpu_flags//_flag/}
 [flags]
 tired            * Adding this option will not query nvidia-smi if gpu is in suspend mode 
 startup          * Useful if you want a certain GPU to be set at startup
+no_emoji         * Use glyphs instead of emoji
 
 * If ${USER} declared env = WLR_DRM_DEVICES on hyprland then use this as the primary GPU
 EOF
 exit
     ;;
 esac
+
+# Define glyphs
+if [[ $NO_EMOJI -eq 1 ]]; then
+    temp_lv="85:, 65:, 45:☁, ❄"
+else
+    temp_lv="85:🌋, 65:🔥, 45:☁️, ❄️"
+fi
+util_lv="90:, 60:󰓅, 30:󰾅, 󰾆" 
 
 nvidia_flag=${nvidia_flag:-0} intel_flag=${intel_flag:-0} amd_flag=${amd_flag:-0}
 #? Based on the flags, call the corresponding function multi flags means multi GPU.