--- a/Configs/.config/waybar/modules/cpuinfo.jsonc
+++ b/Configs/.config/waybar/modules/cpuinfo.jsonc
@@ -2,7 +2,7 @@
         "exec": " ~/.config/hypr/scripts/cpuinfo.sh",
         "return-type": "json",
         "format": "{}",
-        "interval": 5, // once every 5 seconds
+        "restart-interval": 5, // once every 5 seconds
         "tooltip": true,
         "max-length": 1000
     },