--- a/Configs/.config/waybar/modules/cpuinfo.jsonc
+++ b/Configs/.config/waybar/modules/cpuinfo.jsonc
@@ -1,5 +1,5 @@
     "custom/cpuinfo": {
-        "exec": " ~/.config/hypr/scripts/cpuinfo.sh",
+        "exec": "NO_EMOJI=1 ~/.config/hypr/scripts/cpuinfo.sh",
         "return-type": "json",
         "format": "{}",
         "restart-interval": 5, // once every 5 seconds