--- a/Configs/.config/waybar/modules/gpuinfo.jsonc
+++ b/Configs/.config/waybar/modules/gpuinfo.jsonc
@@ -1,5 +1,5 @@
     "custom/gpuinfo": {
-        "exec": " ~/.config/hypr/scripts/gpuinfo.sh",
+        "exec": "NO_EMOJI=1 ~/.config/hypr/scripts/gpuinfo.sh",
         "return-type": "json",
         "format": "{}",
         "interval": 5, // once every 5 seconds
@@ -9,7 +9,7 @@
     },
 
     "custom/gpuinfo#nvidia": {
-        "exec": " ~/.config/hypr/scripts/gpuinfo.sh --use nvidia ",
+        "exec": "NO_EMOJI=1 ~/.config/hypr/scripts/gpuinfo.sh --use nvidia ",
         "return-type": "json",
         "format": "{}",
         "interval": 5, // once every 5 seconds
@@ -18,7 +18,7 @@
     },
 
     "custom/gpuinfo#amd": {
-        "exec": " ~/.config/hypr/scripts/gpuinfo.sh --use amd ",
+        "exec": "NO_EMOJI=1 ~/.config/hypr/scripts/gpuinfo.sh --use amd ",
         "return-type": "json",
         "format": "{}",
         "interval": 5, // once every 5 seconds
@@ -27,7 +27,7 @@
     },
 
     "custom/gpuinfo#intel": {
-        "exec": " ~/.config/hypr/scripts/gpuinfo.sh --use intel ",
+        "exec": "NO_EMOJI=1 ~/.config/hypr/scripts/gpuinfo.sh --use intel ",
         "return-type": "json",
         "format": "{}",
         "interval": 5, // once every 5 seconds