## Icon generation bottleneck

The function to generate icons have been causing performance bottleneck since the beginning of cpuinfo module.  The original version (3cf5d2971ba43a71c95bde8c3ec67858dfa12ed2, `c.sh` in the screenshot) has been taking around 30.5ms (not absolute, for comparison).

To reduce the performance overhead, the revision (9b9b1ce8cdd5914f243f62f19d38c6c2d457fd8b, `d.sh` in the screenshot) have improved the speed to 22.3ms (**37% faster** than the original version). Here a single `jq` has been used to replace multiple forks of `awk` to improve the performance. 

![image](https://github.com/prasanthrangan/hyprdots/assets/76743829/cfd66289-e841-4eb2-a275-8f8c8861b82d)

The original version turns out to run on 6.58ms (**364% faster** than original and **238% faster** than the revised) after minimizing the number of forks and removing all forks to `awk`  (`b.sh` in the screenshot).  I also tested possibly the minimal version `a.sh` which contains only if branches completed in 5.4ms.

![image](https://github.com/prasanthrangan/hyprdots/assets/76743829/59133eff-f73c-480f-9639-afc4601bc5bf)

Despite `a.sh` is the fastest, I included `b.sh` in both gpuinfo and cpuinfo as it follows Object Oriented principles and performance difference is insignificant (1.18ms). 

## The forks

There are a number of redundant forks throughout script like:
`maxfreq=$(lscpu | grep "CPU max MHz" | awk -F: '{ print $2}' | sed -e 's/ //g' -e 's/\.[0-9]*//g')`

Which could be improved by substituting with:
`maxfreq=$(lscpu | awk '/CPU max MHz/ { sub(/\..*/,"",$4); print $4}')`

The number of forks have been minimized completely on `cpuinfo.sh` and partially on `gpuinfo.sh`.

## `top` bottleneck

The overall performance of the script wasn't improved significantly even after minimizing forks as the `top -bn1` taking almost **260ms** as it has to compute information of all running and sleeping processes.  The utility `mpstat` is a great alternative but it requires the program to be installed on the system. Hence `top` is replaced by calculating the utilization levels directly from `/proc/stat` which is taking around **15ms**, that is **1633%** improvement in performance.

## Continuous script

As the script has to recalculate values that doesn't change during runtime (like CPU name and max freq), it is inefficient to run on every 5s. Since waybar allows the scripts to be executed in continuous mode which allows significant improvement on performance as the static values doesn't required to be recalculated on each iteration, the cpuinfo.sh has been adjusted to run as a continuous script.

## Other changes

Improved accuracy of utilization level as it now prints average between each updates. Also added option to decide whether to print emojis or glyphs by setting an environment variable `NO_EMOJI` as the commit: 9b9b1ce8cdd5914f243f62f19d38c6c2d457fd8b converted emojis to glyphs.

## Warning to reviewers

The changes included in the commit: a5cfdd0f3f5a94f8dfa0ef4fc27dffbc973a6097 have been included in this PR, but haven't tested as my machine doesn't have AMD CPU. Although, I tried testing with passing faked sensors data and worked well.

## Testing

All the unit testing has been done by the script:

`python -m timeit -n 10 -s 'import subprocess' "subprocess.call('./test.sh', shell=True)"`

The `python timeit` is not absolute but is highly precise and useful for comparison. I think `strace` or `time` gives more accurate results but are not statistical or precise.

## Type of change

- [x] **Performance improvement** (non-breaking change)
- [ ] **Bug fix** (non-breaking change which fixes an issue)
- [ ] **New feature** (non-breaking change which adds functionality)
- [ ] **Breaking change** (fix or feature that would cause existing functionality to not work as expected)
- [ ] **Documentation update** (non-breaking change; modified files are limited to the documentations)
- [ ] **Technical debt** (a code change that does not fix a bug or add a feature but makes something clearer for devs)

## Checklist

- [x] I have read the [CONTRIBUTING](https://github.com/prasanthrangan/hyprdots/blob/main/CONTRIBUTING.md) document.
- [x] My code follows the code style of this project.
- [x] My commit message follows the [commit guidelines](https://github.com/prasanthrangan/hyprdots/blob/main/CONTRIBUTING.md#git-commit-messages).
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added a changelog entry.
- [x] I have added necessary comments/documentation to my code.
- [ ] I have added tests to cover my changes.
- [x] I have tested my code locally and it works as expected.
- [ ] All new and existing tests passed.