@kRHYME7 2024-02-24
![image](https://github.com/prasanthrangan/hyprdots/assets/53417443/ac9de7b8-cf7f-46e5-9537-d3aeb9bcac27)
The same applies to the CPU. I think I got missing values here. 

Disclaimer: I did not fully review it yet. Looks good in the description, so I won't dig deeper if this works on my end. 😁

Additionally, I believe you can pass the `$NO_EMOJI` variable to the `query()` function to avoid making major changes to the JSONC. In the `query()` function, there's a file `${gpuQ}` which lists all variables on the script's first execution. You can append the `$NO_EMOJI` variable to this list, and it might be invoked by running the script once at startup, like so:
```
${0} --emoji
```