@mislah 2024-02-24
Oh no, The same happening on my machine too O_O

Demn, I forgot to pass the temperature and utilization value to the function.
Wait fixing..