@mislah 2024-02-24
> Additionally, I believe you can pass the `$NO_EMOJI` variable to the `query()` function to avoid making major changes to the JSONC. In the `query()` function, there's a file `${gpuQ}` which lists all variables on the script's first execution. You can append the `$NO_EMOJI` variable to this list, and it might be invoked by running the script once at startup

I though of adding `NO_EMOJI` as a env var to make it easier for end users to switch between emoji and glyphs as they can easily see those in waybar module templates. I don't think anyone will notice this option if we add the variable in `gpuinfo.sh` as the script is already scaryyy.