@kRHYME7 2024-02-24
![image](https://github.com/prasanthrangan/hyprdots/assets/53417443/6ba15e33-e9ad-470e-b3d1-e6d049afc82d)
 
I meant like this. 
You can add a flag for no_emoji. 

Sorry for making that slightly scary, it's the only way I can think of catering all 3 GPUs.