@mislah 2024-02-24
Umm.. I see. It looks great but feels a bit complicated to me . Can you add that to the script if you don't mind or I can try later today.

> Sorry for making that slightly scary, it's the only way I can think of catering all 3 GPUs.

No no, don't get me wrong, I was saying shell scripts are scary as a whole.

I fixed the bug you pointed out btw.. please check again.