@kRHYME7 2024-02-24
It's just a few lines,  Can I just provide you a gist? As forking a repo takes time. 

Also we have the ``` --reset ```  so we can override/change the "variables"

Should look like this:
![image](https://github.com/prasanthrangan/hyprdots/assets/53417443/b114bbfd-b3d0-4a71-a3fe-da1b8d2af78d)